[CLEANUP] Modified directory structure for future development [CLEANUP] Moved pages, components into separate files
