+++ src/components/App/App.jsx
@@ -1,0 +1,12 @@
+import Header from "../../pages/Header/Header";
+import Main from "../../pages/Main/Main";
+import React from "react";
+
+const App = () => (
+    <div>
+        <Header />
+        <Main />
+    </div>
+);
+
+export default App;
