+++ src/pages/Main/Main.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+const Main = () => {
+    return (
+        <div className="Header">
+            <main>
+                <a href="https://randallsewell.me/blog/index.php" target='_blank'>lonelyBKLYN: Speak!</a>
+            </main>
+        </div>
+    );
+};
+
+export default Main;
